requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
celery==5.3.6
motor==3.3.2
Beanie==1.25.0
python-dotenv==1.0.1
fastapi-users[beanie]==12.1.2